meal_order_app/requirements.txt
# To ensure app dependencies are ported from your virtual environment/host machine into your container, run 'pip freeze > requirements.txt' in the terminal to overwrite this file
flask==2.0.1
gunicorn==20.0.4
SQLAlchemy==1.3.12
psycopg2-binary==2.9.1
python-decouple==3.4